+++ src/main/kotlin/com/c0x12c/featureflag/entity/FeatureFlag.kt
@@ -1,5 +1,6 @@
 package com.c0x12c.featureflag.entity
 
+import com.c0x12c.featureflag.models.MetadataContent
 import com.c0x12c.featureflag.serializer.InstantSerializer
 import com.c0x12c.featureflag.serializer.UuidSerializer
 import kotlinx.serialization.Serializable
@@ -14,7 +15,7 @@ data class FeatureFlag(
   val code: String,
   val description: String? = null,
   val enabled: Boolean = false,
-  val metadata: String = "{}", // JSON String
+  val metadata: MetadataContent? = null,
   @Serializable(with = InstantSerializer::class)
   val createdAt: Instant = Instant.now(),
   @Serializable(with = InstantSerializer::class)