+++ src/main/kotlin/com/c0x12c/featureflag/table/FeatureFlagTable.kt
@@ -9,8 +9,14 @@ object FeatureFlagTable : UUIDTable("feature_flags") {
   val code = varchar("code", 50)
   val description = text("description").nullable()
   val enabled = bool("enabled")
-  val metadata = text("metadata") // Store JSON as text
+  val metadata = text("metadata").nullable() // Store serialized MetadataContent as JSON
+  val metadataType = varchar("metadata_type", 50).nullable() // Store the type of metadata
   val createdAt = timestamp("created_at").default(Instant.now())
   val updatedAt = timestamp("updated_at").nullable()
   val deletedAt = timestamp("deleted_at").nullable()
+
+  init {
+    index(true, metadata)
+    index(true, metadataType)
+  }
 }